2-Projects/Agents/auto-freelance-swarm/requirements.txt
# Auto-Freelance Swarm 依赖
# Python 3.10+

# 核心框架
openai>=1.0.0
langchain>=0.2.0
langgraph>=0.1.0
langchain-core>=0.2.0

# Web框架
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.30.0

# 数据处理
pyyaml>=6.0
python-dotenv>=1.0.0

# 浏览器自动化 (可选)
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 向量数据库 (可选)
chromadb>=0.4.0

# 工具库
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0
tiktoken>=0.5.0

# 代码质量
flake8>=6.0.0
black>=23.0.0
pylint>=3.0.0

# 日志
loguru>=0.7.0

# 其他
python-dateutil>=2.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
